feat!: Correct namespace for pages to uk.gov.android.ui.pages
