--- pages/src/androidTest/java/uk/gov/documentchecking/pages/calculator/RadioChoiceQuestionTest.kt
@@ -1,79 +1,0 @@
-package uk.gov.documentchecking.pages.calculator
-
-import android.content.Context
-import androidx.annotation.ArrayRes
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.ComposeContentTestRule
-import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.performScrollTo
-import androidx.test.core.app.ApplicationProvider
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.Parameterized
-
-@RunWith(Parameterized::class)
-class RadioChoiceQuestionTest(
-    private val parameters: RadioChoiceQuestionParameters
-) {
-    private val context: Context = ApplicationProvider.getApplicationContext()
-
-    private val resources = context.resources
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @Test
-    fun informationScreenTests() {
-        composeTestRule.apply {
-            setContent {
-                RadioChoiceQuestion(parameters)
-            }
-
-            onNodeWithText(
-                resources.getString(parameters.title)
-            ).apply {
-                assertIsDisplayed()
-            }
-
-            parameters.content?.let { checkContentSection(this, it) }
-
-            parameters.radioOptions.forEach {
-                onNodeWithText(
-                    it.text
-                ).apply {
-                    performScrollTo()
-                    assertIsDisplayed()
-                }
-            }
-
-            onNodeWithText(
-                resources.getString(parameters.primaryButtonText)
-            ).apply {
-                performScrollTo()
-                assertIsDisplayed()
-            }
-        }
-    }
-
-    private fun checkContentSection(
-        tesRule: ComposeContentTestRule,
-        @ArrayRes
-        text: Int
-    ) {
-        tesRule.apply {
-            resources.getStringArray(text).forEach {
-                onNodeWithText(it).apply {
-                    assertIsDisplayed()
-                }
-            }
-        }
-    }
-
-    companion object {
-        @JvmStatic
-        @Parameterized.Parameters
-        fun values() = RadioChoiceQuestionProvider().values.toList()
-    }
-}
