+++ pages/src/main/java/uk/gov/android/ui/pages/brp/BrpInstructionsContentSection.kt
@@ -1,0 +1,11 @@
+package uk.gov.android.ui.pages.brp
+
+import androidx.annotation.ArrayRes
+import androidx.annotation.StringRes
+
+data class BrpInstructionsContentSection(
+    @StringRes
+    val subTitle: Int? = null,
+    @ArrayRes
+    val text: Int
+)
