+++ pages/src/main/java/uk/gov/android/ui/pages/errors/ErrorPageTags.kt
@@ -1,0 +1,8 @@
+package uk.gov.android.ui.pages.errors
+
+object ErrorPageTags {
+    const val information = "Information"
+    const val primaryButton = "PrimaryButton"
+    const val rootLayout = "ErrorPage"
+    const val secondaryButton = "SecondaryButton"
+}
