--- pages/src/main/java/uk/gov/documentchecking/pages/brp/BrpInstructionsContentSection.kt
@@ -1,11 +1,0 @@
-package uk.gov.documentchecking.pages.brp
-
-import androidx.annotation.ArrayRes
-import androidx.annotation.StringRes
-
-data class BrpInstructionsContentSection(
-    @StringRes
-    val subTitle: Int? = null,
-    @ArrayRes
-    val text: Int
-)
