--- pages/src/main/java/uk/gov/documentchecking/pages/errors/ErrorPageTags.kt
@@ -1,8 +1,0 @@
-package uk.gov.documentchecking.pages.errors
-
-object ErrorPageTags {
-    const val information = "Information"
-    const val primaryButton = "PrimaryButton"
-    const val rootLayout = "ErrorPage"
-    const val secondaryButton = "SecondaryButton"
-}
