--- pages/src/test/java/uk/gov/documentchecking/pages/calculator/NumberInputQuestionScreenshotTest.kt
@@ -1,31 +1,0 @@
-package uk.gov.documentchecking.pages.calculator
-
-import androidx.compose.runtime.Composable
-import com.android.resources.NightMode
-import org.junit.runner.RunWith
-import org.junit.runners.Parameterized
-import org.junit.runners.Parameterized.Parameters
-import uk.gov.documentchecking.pages.BaseScreenshotTest
-
-@RunWith(Parameterized::class)
-class NumberInputQuestionScreenshotTest(
-    private val parameters: Pair<NumberInputQuestionParameters, NightMode>
-) : BaseScreenshotTest(parameters.second) {
-
-    override val generateComposeLayout: @Composable () -> Unit = {
-        NumberInputQuestion(parameters.first)
-    }
-
-    companion object {
-        @JvmStatic
-        @Parameters
-        fun values(): List<Pair<NumberInputQuestionParameters, NightMode>> {
-            val result: MutableList<Pair<NumberInputQuestionParameters, NightMode>> =
-                mutableListOf()
-
-            NumberInputQuestionProvider().values.forEach(applyNightMode(result))
-
-            return result
-        }
-    }
-}
