--- pages/src/test/java/uk/gov/documentchecking/pages/errors/ErrorPageScreenshotTest.kt
@@ -1,30 +1,0 @@
-package uk.gov.documentchecking.pages.errors
-
-import androidx.compose.runtime.Composable
-import com.android.resources.NightMode
-import org.junit.runner.RunWith
-import org.junit.runners.Parameterized
-import org.junit.runners.Parameterized.Parameters
-import uk.gov.documentchecking.pages.BaseScreenshotTest
-
-@RunWith(Parameterized::class)
-class ErrorPageScreenshotTest(
-    private val parameters: Pair<ErrorPageParameters, NightMode>
-) : BaseScreenshotTest(parameters.second) {
-
-    override val generateComposeLayout: @Composable () -> Unit = {
-        ErrorPage(parameters.first)
-    }
-
-    companion object {
-        @JvmStatic
-        @Parameters
-        fun values(): List<Pair<ErrorPageParameters, NightMode>> {
-            val result: MutableList<Pair<ErrorPageParameters, NightMode>> = mutableListOf()
-
-            ErrorPageProvider().values.forEach(applyNightMode(result))
-
-            return result
-        }
-    }
-}
